chore(list): add and use size.js utility
